Попытка разделить на модули.Error
