+++ js/main.js
@@ -1,3 +1,5 @@
+import { getRandomArrayElement, getGeneraRandomId } from './util.js';
+
 const PHOTOS_OBJECTS_COUNT = 25;
 const LIKES_MIN_VALUE = 15;
 const LIKES_MAX_VALUE = 200;
@@ -50,26 +52,7 @@ const NAMES_COMMENTS = [
   'Антон',
 ];
 
-//Получение уникальных ID в заданном диапазоне
-const getGeneraRandomId = (min, max) => {
-  /*
-    Math.ceil   - Возвращает наименьшее целое число, большее или равное его числовому аргументу.
-    Math.min    - Возвращает меньшее из набора предоставленных числовых выражений.
-    Math.max    - Возвращает большее из набора предоставленных числовых выражений.
-    Math.abs    - Возвращает абсолютное значение числа.
-    Math.random - Возвращает случайное число от 0 (вкл) до 1 (не вкл).
-    Math.floor  - Округляет до ближайшего меньшего.
-  */
-  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
-  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-//Поиск случайного элемента в переданном массиве. (декомпозиция)
-const getRandomArrayElement = (array) => array[getGeneraRandomId(0, array.length - 1)];
-
-// Получение номера комментария
+// Получение номера комментария (замыкание)
 const getRandomIdComments = () => {
   let lastGaneratedId = 0;
   return () => {
